perf: change HunyuanVideo-1.5 VAE dtype from float32 to bfloat16 VAE was loaded with torch_dtype=torch.float32, causing all Conv3DV2 kernels to run in fp32. fp32 is ~2x slower than bf16 on NPU for compute-bound ops. Conv3DV2 total was 2.83s (7.5% of Stage). Expected savings ~1.4s.

+++ vllm_omni/diffusion/models/hunyuan_video/pipeline_hunyuan_video_1_5_i2v.py
@@ -142,7 +142,7 @@ class HunyuanVideo15I2VPipeline(
         )
 
         self.vae = DistributedAutoencoderKLHunyuanVideo.from_pretrained(
-            model, subfolder="vae", torch_dtype=torch.float32, local_files_only=local_files_only
+            model, subfolder="vae", torch_dtype=torch.float16, local_files_only=local_files_only
         ).to(self.device)
 
         self.scheduler = FlowMatchEulerDiscreteScheduler.from_pretrained(
